Fix AB1
This now passes the output data id to the AB1 measurement code.
That allows the measurement to filter out bands not involved in
the comparison which wasn't happening before.  This means that
the runQuantum method was overridden to pass that information
down.  It means any measurement retargetted to the MatchMulti
measurement task must take two id arguments.