Change how we skip the stellar locust test
We can't both auto scan to lint our code and use a list of tests to skip.
If we want to do both, we have make pytest ignore the test by renaming it.
Only files that start with test_ get run by pytest, so renaming the file will do the trick