Accommodate WCS to None on fail instead of raise
In single frame processing, instead of raising, we now set the WCS of an
exposure to None if the astrometry fails to find a solution and allow
processing to proceed.  As such, some accommodations must be made for
downstream tasks and afterburners that require the presence of a WCS to
be meaningful.  Here, we add an option to the base measure class as to
whether valid astrometric and/or photometric solutions are required.  We
default this to True as most metrics do require such calibrations (indeed,
many involve cross-epoch matching, but require WCS information), but some
conceivably do not (e.g. straight detection number counts).  When these
config parameters are True, we skip over dataIds with no valid WCS and/or
photoCalib entries, thus excluding them in the metrics measurements.

Various permutations of these config settings and presence of WCS and
photoCalibs have been added to the tests.